+++ tests/compiler/test_tracing.py
@@ -145,3 +145,30 @@ D()
     (types, _) = compiler.do_compile()
     files = types["__config__::C"].get_all_instances()
     assert len(files) == 1
+
+
+def test_747_entity_multi_location(snippetcompiler):
+    snippetcompiler.setup_for_snippet("""
+entity Alpha:
+    string name
+end
+
+implementation none for Alpha:
+end
+implement Alpha using none
+
+index Alpha(name)
+
+a= Alpha(name="A")
+b= Alpha(name="A")
+c= Alpha(name="A")
+""", autostd=False)
+    (_, scopes) = compiler.do_compile()
+
+    root = scopes.get_child("__config__")
+    a = root.lookup("a").get_value()
+    assert len(a.get_locations()) == 3
+    assert sorted([l.lnr for l in a.get_locations()]) == [12, 13, 14]
+
+
+
